src/requirements.txt
Flask==1.1.2
google-cloud-storage
google-cloud-dlp
google-cloud-pubsub
gunicorn==20.0.4